Refactor from PR discussion
